Fullscreen Renderer feature
Example to illustrate how to create a full-screen effect. Therefore, I created 3 shaders to show how this function works and render the scene like a Gameboy, in grayscale and inverting the colors.

Additionally, I changed the name of the Blur feature and pass, for a better name-matching with the other features.